% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsw.R
\name{fsw.ivreg}
\alias{fsw.ivreg}
\title{Calculate the conditional F-statistic of Sanderson and Windmeijer (2016)}
\usage{
\method{fsw}{ivreg}(mod)
}
\arguments{
\item{mod}{A fitted model from \code{ivreg::ivreg()}, i.e. an object of class ivreg.}
}
\value{
An object of class fsw with the following elements:
\itemize{
\item fsw vector of conditional \emph{F}-statistics
}
}
\description{
Calculate the conditional F-statistic of Sanderson and Windmeijer (2016)
}
\examples{
require(ivreg)
set.seed(12345)
n <- 4000
z1 <- rnorm(n)
z2 <- rnorm(n)
w1 <- rnorm(n)
w2 <- rnorm(n)
u <- rnorm(n)
x1 <- z1 + z2 + 0.2*u + 0.1*w1 + rnorm(n)
x2 <- z1 + 0.94*z2 - 0.3*u + 0.1*w2 + rnorm(n)
y <- x1 + x2 + w1 + w2 + u
dat <- data.frame(w1, w2, x1, x2, y, z1, z2)
mod <- ivreg::ivreg(y ~ x1 + x2 + w1 + w2 | z1 + z2 + w1 + w2, data = dat)
fsw(mod)

}
\references{
Sanderson E and Windmeijer F. A weak instrument \emph{F}-test in linear
IV models with multiple endogenous variables. Journal of Econometrics,
2016, 190, 2, 212-221, \doi{10.1016/j.jeconom.2015.06.004}.
}
