% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmm.R
\name{lsmm}
\alias{lsmm}
\title{Logistic structural mean model}
\usage{
lsmm(
  formula,
  instruments,
  data,
  subset,
  na.action,
  contrasts = NULL,
  t0 = NULL,
  ...
)
}
\description{
The logistic structural mean model is due to Robins ().
\code{lsmm} implements
}
