% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmm.R
\name{msmm}
\alias{msmm}
\title{Multiplicative structural mean model}
\usage{
msmm(
  formula,
  instruments,
  data,
  subset,
  na.action,
  weights,
  offset,
  contrasts = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  estmethod = c("gmm", "gmmalt", "tsls", "tslsalt"),
  ...
)
}
\arguments{
\item{formula}{The model formula}

\item{estmethod}{Estimation method, these are
\itemize{
\item \code{"gmm"} GMM estimation of the MSMM
\item \code{"gmmalt"} GMM estimation of the alternative moment conditions
for the MSMM as per Clarke et al. (2015)
\item \code{"tsls"} the TSLS method of fitting MSMM Clarke et al. (2015)
\item \code{"tslsalt"} the alternative TSLS method of Clarke et al. (2015)
}}
}
\description{
Function providing several methods to estimate the multiplicative
structural mean model (MSMM) of Robins (1994).
}
\examples{
# Data generation from the example in the [`ivtools::ivglm()`] helpfile
set.seed(9)
n <- 1000
psi0 <- 0.5
psi1 <- 0.2
Z <- rbinom(n, 1, 0.5)
X <- rbinom(n, 1, 0.7*Z + 0.2*(1 - Z))
m0 <- plogis(1 + 0.8*X - 0.39*Z)
Y <- rbinom(n, 1, plogis(psi0*X + log(m0/(1 - m0))))
dat <- data.frame(Z, X, Y)
msmm(Y ~ X | Z, data = dat, estmethod = "tsls")
}
\references{
Clarke PS, Palmer TM Windmeijer F. Estimating structural
mean models with multiple instrumental variables using the
Generalised Method of Moments. Statistical Science, 2015, 30, 1,
96-117. \doi{10.1214/14-STS503}

Hernan and Robins. Instruments for causal inference: An
epidemiologist's dream? Epidemiology, 2006, 17, 360-372.
\doi{10.1097/01.ede.0000222409.00878.37}

Robins JM. ROBINS, J. M. (1989). The analysis of randomised and
nonrandomised AIDS treatment trials using a new approach to
causal inference in longitudinal studies.
In Health Service Research Methodology: A Focus on AIDS
(L. Sechrest, H. Freeman and A. Mulley, eds.) 113–159.
US Public Health Service, National Center for Health Services Research,
Washington, DC.
}
