% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesamplemr-package.R
\docType{package}
\name{OneSampleMR-package}
\alias{OneSampleMR-package}
\alias{onesamplemr}
\alias{OneSampleMR}
\title{OneSampleMR: Use functions for one-sample Mendelian randomization
and instrumental variables analyses}
\description{
Some useful functions for one-sample Mendelian randomization
and instrumental variables analyses.
}
