% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmm.R
\name{summary.lsmm}
\alias{summary.lsmm}
\alias{print.tsps}
\alias{print.summary.lsmm}
\title{Summarizing LSMM Fits}
\usage{
\method{summary}{lsmm}(object, ...)

\method{print}{tsps}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{summary.lsmm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"lsmm"}.}

\item{...}{further arguments passed to or from other methods.

S3 summary and print methods for objects of class \code{lsmm} and \code{summary.lsmm}.}

\item{x}{an object of class \code{"summary.lsmm"}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
\code{summary.lsmm()} returns an object of class \code{"summary.lsmm"}. A list with the following elements:

\item{smry}{An object from a call to \code{\link[gmm:summary]{gmm::summary.gmm()}}}
\item{object}{The object of class \code{tsps} passed to the function.}
}
\description{
Summarizing LSMM Fits
}
\examples{
# For examples see the examples at the bottom of help('msmm')
}
