% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmm.R
\name{summary.msmm}
\alias{summary.msmm}
\alias{print.msmm}
\alias{print.summary.msmm}
\title{Summarizing MSMM Fits}
\usage{
\method{summary}{msmm}(object, ...)

\method{print}{msmm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{summary.msmm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"msmm"}.}

\item{...}{further arguments passed to or from other methods.

S3 summary and print methods for objects of class \code{msmm} and \code{summary.msmm}.}

\item{x}{an object of class \code{"summary.msmm"}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
\code{summary.msmm()} returns an object of class \code{"summary.msmm"}. A list with the following elements:

\item{smry}{An object from a call to either \code{\link[gmm:summary]{gmm::summary.gmm()}} or
\code{\link[ivreg:summary.ivreg]{ivreg::summary.ivreg()}}.}
\item{object}{The object of class \code{msmm} passed to the function.}
}
\description{
Summarizing MSMM Fits
}
\examples{
# For examples see the examples at the bottom of help('msmm')
}
