% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsps.R
\name{summary.tsps}
\alias{summary.tsps}
\alias{print.tsps}
\alias{print.summary.tsps}
\title{Summarizing TSPS Fits}
\usage{
\method{summary}{tsps}(object, ...)

\method{print}{tsps}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{summary.tsps}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"tsps"}.}

\item{...}{further arguments passed to or from other methods.

S3 summary and print methods for objects of class \code{tsps} and \code{summary.tsps}.}

\item{x}{an object of class \code{"summary.tsps"}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
\code{summary.tsps()} returns an object of class \code{"summary.tsps"}. A list with the following elements:

\item{smry}{An object from a call to either \code{\link[gmm:summary]{gmm::summary.gmm()}}
\item{object}{The object of class \code{tsps} passed to the function.}
}
\description{
Summarizing TSPS Fits
}
\examples{
# For examples see the examples at the bottom of help('tsps')
}
