% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsri.R
\name{summary.tsri}
\alias{summary.tsri}
\alias{print.tsri}
\alias{print.summary.tsri}
\title{Summarizing TSRI Fits}
\usage{
\method{summary}{tsri}(object, ...)

\method{print}{tsri}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{summary.tsri}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"tsri"}.}

\item{...}{further arguments passed to or from other methods.

S3 summary and print methods for objects of class \code{tsri} and \code{summary.tsri}.}

\item{x}{an object of class \code{"summary.tsri"}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
\code{summary.tsri()} returns an object of class \code{"summary.tsri"}. A list with the following elements:

\item{smry}{An object from a call to \code{\link[gmm:summary]{gmm::summary.gmm()}}}
\item{object}{The object of class \code{tsps} passed to the function.}
}
\description{
Summarizing TSRI Fits
}
\examples{
# For examples see the examples at the bottom of help('tsri')
}
