% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsps.R
\name{tsps}
\alias{tsps}
\title{Two-stage predictor substitution (TSPS) estimators}
\usage{
tsps(
  formula,
  instruments,
  data,
  subset,
  na.action,
  contrasts = NULL,
  t0 = NULL,
  ...
)
}
\description{
TSPS estimators
\code{tsps} implements
}
