% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsri.R
\name{tsri}
\alias{tsri}
\title{Two-stage residual inclusion (TSRI) estimators}
\usage{
tsri(
  formula,
  instruments,
  data,
  subset,
  na.action,
  contrasts = NULL,
  t0 = NULL,
  ...
)
}
\description{
TSRI estimators
\code{tsri} implements
}
