% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsw.R
\name{fsw}
\alias{fsw}
\title{Conditional F-statistic of Sanderson and Windmeijer}
\usage{
fsw(object)
}
\arguments{
\item{object}{An object containing the results of an IV model for which to
calculate the conditional F-statistics for each endogenous variable.}
}
\description{
\code{fsw} is a generic function for calculating the conditional F-statistic of
Sanderson and Windmeijer (2016). The function invokes particular methods which
depend upon the \code{\link{class}} of the first argument.
}
\seealso{
\code{\link[=fsw.ivreg]{fsw.ivreg()}}
}
