% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmm.R
\name{lsmm}
\alias{lsmm}
\title{Logistic structural mean model}
\usage{
lsmm(
  formula,
  instruments,
  data,
  subset,
  na.action,
  contrasts = NULL,
  t0 = NULL,
  ...
)
}
\arguments{
\item{formula}{formula specification(s) of the regression
relationship and the instruments. Either \code{instruments} is missing and
\code{formula} has three parts as in \code{y ~ x1 + x2 | z1 + z2 + z3}
(recommended) or \code{formula} is \code{y ~ x1 + x2} and \code{instruments}
is a one-sided formula \code{~ z1 + z2 + z3} (only for backward
compatibility).}

\item{instruments}{formula specification(s) of the regression
relationship and the instruments. Either \code{instruments} is missing and
\code{formula} has three parts as in \code{y ~ x1 + x2 | z1 + z2 + z3}
(recommended) or \code{formula} is \code{y ~ x1 + x2} and \code{instruments}
is a one-sided formula \code{~ z1 + z2 + z3} (only for backward
compatibility).}

\item{data}{an optional data frame containing the variables in the model.
By default the variables are taken from the environment of the
\code{formula}.}

\item{subset}{an optional vector specifying a subset of observations to be
used in fitting the model.}

\item{na.action}{a function that indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} option.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of
\code{\link[stats:model.matrix]{model.matrix.default}}.}

\item{t0}{A vector of starting values for the gmm optimizer. This should
have length equal to the number of exposures plus 1.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The logistic structural mean model (LSMM) is due to Robins (1989).
This command implements the double-LSMM due to Vansteelandt and
Goetghebeur (2003) using the GMM estimation approach of Clarke et al. (2015).
}
\references{
Clarke PS, Palmer TM, Windmeijer F. Estimating structural
mean models with multiple instrumental variables using the
Generalised Method of Moments. Statistical Science, 2015, 30, 1,
96-117. \doi{10.1214/14-STS503}

Robins JM. The analysis of randomised and
nonrandomised AIDS treatment trials using a new approach to
causal inference in longitudinal studies.
In Health Service Research Methodology: A Focus on AIDS
(L. Sechrest, H. Freeman and A. Mulley, eds.). 1989. 113–159.
US Public Health Service, National Center for Health Services Research,
Washington, DC.

Vansteelandt S, Geotghebeur E.
Causal inference with generalized structural mean models.
Journal of the Royal Statistical Society (Series B).
2003, 65, 4, 818-835.
\doi{10.1046/j.1369-7412.2003.00417.x}
}
