% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesamplemr-package.R
\docType{package}
\name{OneSampleMR-package}
\alias{OneSampleMR-package}
\alias{onesamplemr}
\alias{OneSampleMR}
\title{OneSampleMR: One Sample Mendelian Randomization and Instrumental Variable Analyses}
\description{
Useful functions for one-sample (individual level data) Mendelian randomization and instrumental variable analyses. The package includes implementations of; the Sanderson and Windmeijer (2016) \doi{10.1016/j.jeconom.2015.06.004} conditional F-statistic, the multiplicative structural mean model Hernán and Robins (2006) \doi{10.1097/01.ede.0000222409.00878.37}, and two-stage predictor substitution and two-stage residual inclusion estimators explained by Terza et al. (2008) \doi{10.1016/j.jhealeco.2007.09.009}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/remlapmot/OneSampleMR}
  \item \url{https://remlapmot.github.io/OneSampleMR/}
  \item \url{https://mrcieu.r-universe.dev/OneSampleMR}
  \item Report bugs at \url{https://github.com/remlapmot/OneSampleMR/issues/}
}

}
\author{
\strong{Maintainer}: Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})

Authors:
\itemize{
  \item Wes Spiller \email{wes.spiller@bristol.ac.uk} (\href{https://orcid.org/0000-0002-8169-5531}{ORCID})
  \item Eleanor Sanderson \email{eleanor.sanderson@bristol.ac.uk} (\href{https://orcid.org/0000-0001-5188-5775}{ORCID})
}

}
